requirements.txt
# Web / API
fastapi>=0.110
uvicorn[standard]>=0.27
watchfiles>=0.21

# Pydantic v2 + Settings
pydantic>=2.7
pydantic-settings>=2.2

# DB
SQLAlchemy>=2.0
psycopg[binary]>=3.1

# HTTP
requests>=2.32
httpx>=0.27

# Utilidades
python-dateutil>=2.9

# Auth
python-jose[cryptography]>=3.3
passlib[bcrypt]==1.7.4
email-validator>=2.1
bcrypt==4.0.1

deep-translator
pytest
pytest-cov
httpx
freezegun